requirements.txt
python-dotenv
djangorestframework
psycopg2-binary
supabase
django-environ
djangorestframework-simplejwt
django-cors-headers
redis
django-redis
upstash-redis
gunicorn
dj-database-url
django-anymail[resend]
pyjwt
cryptography
razorpay
celery[redis]>=5.3
redis[hiredis]>=5.0
django-celery-results